Move schemas to json
