+++ messages/es.json
@@ -9,5 +9,15 @@
   "admin/editor.condition-layout.subject": "Tema",
   "admin/editor.condition-layout.verb": "Verbo (operador)",
   "admin/editor.condition-layout.object": "Objeto",
-  "admin/editor.condition-layout.match": "Tipo de match"
+  "admin/editor.condition-layout.match": "Tipo de match",
+  "admin/editor.condition-layout.product.subject.productId": "Product ID",
+  "admin/editor.condition-layout.product.subject.categoryId": "Category ID",
+  "admin/editor.condition-layout.product.subject.brandId": "Brand ID",
+  "admin/editor.condition-layout.product.subject.productClusters": "Product clusters",
+  "admin/editor.condition-layout.product.subject.categoryTree": "Category tree",
+  "admin/editor.condition-layout.product.subject.selectedItemId": "Selected Item ID",
+  "admin/editor.condition-layout.verb.is": "is",
+  "admin/editor.condition-layout.verb.is-not": "is not",
+  "admin/editor.condition-layout.verb.contains": "contains",
+  "admin/editor.condition-layout.verb.does-not-contain": "does not contain"
 }
