+++ react/Condition.ts
@@ -57,59 +57,4 @@ const Condition: StorefrontFunctionComponent<ConditionProps> = ({
   return (children as any) ?? null
 }
 
-Condition.schema = {
-  title: 'admin/editor.condition-layout.condition',
-  type: 'object',
-  properties: {
-    enabled: {
-      title: 'admin/editor.condition-layout.enabled.title',
-      description: 'admin/editor.condition-layout.enabled.description',
-      default: true,
-      type: 'boolean',
-    },
-  },
-  dependencies: {
-    enabled: {
-      oneOf: [
-        {
-          properties: {
-            enabled: {
-              enum: [true],
-            },
-            match: {
-              default: 'all',
-              enum: ['any', 'none', 'all'],
-              title: 'admin/editor.condition-layout.match',
-              type: 'string',
-            },
-            conditions: {
-              title: 'admin/editor.condition-layout.conditions',
-              type: 'array',
-              items: {
-                type: 'object',
-                properties: {
-                  subject: {
-                    type: 'string',
-                    title: 'admin/editor.condition-layout.subject',
-                  },
-                  verb: {
-                    default: 'is',
-                    type: 'string',
-                    title: 'admin/editor.condition-layout.verb',
-                    enum: ['is', 'is-not', 'contains', 'does-not-contain'],
-                  },
-                  object: {
-                    type: 'string',
-                    title: 'admin/editor.condition-layout.object',
-                  },
-                },
-              },
-            },
-          },
-        },
-      ],
-    },
-  },
-}
-
 export default Condition
