--- react/modules/subjects.ts
@@ -1,22 +1,0 @@
-export const PRODUCT_SUBJECTS = {
-  productId: {
-    type: 'value',
-  },
-  categoryId: {
-    type: 'value',
-  },
-  brandId: {
-    type: 'value',
-  },
-  productClusters: {
-    type: 'array',
-    id: 'id',
-  },
-  categoryTree: {
-    type: 'array',
-    id: 'id',
-  },
-  selectedItemId: {
-    type: 'value',
-  },
-} as const
