+++ store/contentSchemas.json
@@ -1,0 +1,128 @@
+{
+  "definitions": {
+    "ConditionLayoutProduct": {
+      "title": "admin/editor.condition-layout.wrapper.product"
+    },
+    "ConditionProduct": {
+      "title": "admin/editor.condition-layout.condition",
+      "type": "object",
+      "properties": {
+        "enabled": {
+          "title": "admin/editor.condition-layout.enabled.title",
+          "description": "admin/editor.condition-layout.enabled.description",
+          "default": true,
+          "type": "boolean"
+        }
+      },
+      "dependencies": {
+        "enabled": {
+          "oneOf": [
+            {
+              "properties": {
+                "enabled": {
+                  "enum": [true]
+                },
+                "match": {
+                  "default": "all",
+                  "enum": ["any", "none", "all"],
+                  "title": "admin/editor.condition-layout.match",
+                  "type": "string"
+                },
+                "conditions": {
+                  "title": "admin/editor.condition-layout.conditions",
+                  "type": "array",
+                  "items": {
+                    "type": "object",
+                    "properties": {
+                      "subject": {
+                        "default": "",
+                        "type": "string",
+                        "title": "admin/editor.condition-layout.subject",
+                        "enum": [
+                          "productId",
+                          "categoryId",
+                          "brandId",
+                          "productClusters",
+                          "categoryTree",
+                          "selectedItemId"
+                        ],
+                        "enumNames": [
+                          "admin/editor.condition-layout.product.subject.productId",
+                          "admin/editor.condition-layout.product.subject.categoryId",
+                          "admin/editor.condition-layout.product.subject.brandId",
+                          "admin/editor.condition-layout.product.subject.productClusters",
+                          "admin/editor.condition-layout.product.subject.categoryTree",
+                          "admin/editor.condition-layout.product.subject.selectedItemId"
+                        ]
+                      },
+                      "verb": {
+                        "type": "string",
+                        "title": "admin/editor.condition-layout.verb"
+                      },
+                      "object": {
+                        "type": "string",
+                        "title": "admin/editor.condition-layout.object"
+                      }
+                    },
+                    "dependencies": {
+                      "subject": {
+                        "oneOf": [
+                          {
+                            "properties": {
+                              "subject": {
+                                "enum": ["productClusters", "categoryTree"]
+                              },
+                              "verb": {
+                                "default": "contains",
+                                "enum": ["contains", "does-not-contain"],
+                                "enumNames": [
+                                  "admin/editor.condition-layout.verb.contains",
+                                  "admin/editor.condition-layout.verb.does-not-contain"
+                                ]
+                              }
+                            }
+                          },
+                          {
+                            "properties": {
+                              "subject": {
+                                "enum": [
+                                  "productId",
+                                  "categoryId",
+                                  "brandId",
+                                  "selectedItemId"
+                                ]
+                              },
+                              "verb": {
+                                "default": "is",
+                                "enum": ["is", "is-not"],
+                                "enumNames": [
+                                  "admin/editor.condition-layout.verb.is",
+                                  "admin/editor.condition-layout.verb.is-not"
+                                ]
+                              }
+                            }
+                          }
+                        ]
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          ]
+        }
+      }
+    },
+    "ConditionElse": {
+      "title": "admin/editor.condition-layout.else",
+      "type": "object",
+      "properties": {
+        "enabled": {
+          "default": true,
+          "title": "admin/editor.condition-layout.enabled.title",
+          "type": "boolean"
+        }
+      }
+    }
+  }
+}
